@@ -23,7 +23,7 @@ add_code_coverage_all_targets(EXCLUDE odbc-test)
 if (NOT "$ENV{BOOST_ROOT}" STREQUAL "")
     set(BOOST_ROOT "$ENV{BOOST_ROOT}")
 endif()
-set(CMAKE_CXX_STANDARD 98)
+set(CMAKE_CXX_STANDARD 11)
 
 set(CMAKE_PROJECT_VERSION ${PROJECT_VERSION})
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/cmake")
